requirements.txt
autopep8
numpy==1.17.3
opencv-python==4.1.1.26
Pillow==6.2.1
pip==21.1.1
setuptools==39.0.1
